init.py brought out of flake8 checks
