+++ dff/script/core/types.py
@@ -33,7 +33,7 @@ ModuleName = str
 """
 Module name names addon state, or your own module state. For example module name can be `"df_db_connector"`.
 """
-# todo: change example
+# TODO: change example
 
 
 # TODO: add description for each stage of ActorStage
