+++ dff/script/labels/__init__.py
@@ -1,1 +1,1 @@
-from .std_labels import *  # noqa: F401, F403
+from .std_labels import repeat, previous, to_start, to_fallback, forward, backward
