+++ dff/utils/testing/__init__.py
@@ -1,3 +1,3 @@
-from .common import *  # noqa: F401, F403
-from .toy_script import *  # noqa: F401, F403
-from .response_comparers import *  # noqa: F401, F403
+from .common import is_interactive_mode, check_happy_path, run_interactive_mode
+from .toy_script import TOY_SCRIPT, HAPPY_PATH
+from .response_comparers import default_comparer, generics_comparer
